Remote Build Scripts
Added scripts for remotely building and installing plugin from desktop development machine via SSH.

+++ scripts/norns_build_install_offline.sh
@@ -3,17 +3,29 @@
 # Adapted from:
 # https://github.com/madskjeldgaard/portedplugins/blob/main/scripts/rpi_build
 #
-# This version for offline build/install on Norns. Requires SuperCollider source in same dir as plugin project.
-# This project directory and SuperCollider source should be installed at path /home/we/dust/dev (or edit BASE_DIR below).
+# This version for offline build/install on Norns when Norns itself has no internet access.
+# Requires SuperCollider source in same dir as this plugin project.
+# This plugin project directory and SuperCollider source should be installed at path /home/we/dust/development/SuperCollider.
+# See below for required paths.
+# Script to be run locally on Norns via SSH or called remotely run on the device from local development machine
+# (see scripts/norns_build_remote.sh).
 
 # Temp dir for build
-BASE_DIR="/home/we/dev"
+BASE_DIR="/home/we/development/SuperCollider"
 # Project directory
 PROJ_DIR="${BASE_DIR}/Open303_SuperCollider"
 # SuperCollider source directory
 SC_DIR="${BASE_DIR}/supercollider"
+# Intermediate build directory
+INSTALL_CACHE="${PROJ_DIR}/build/install_cache"
 # SuperCollider extension install directory
-INSTALL_DIR=$HOME/.local/share/SuperCollider/Extensions
+INSTALL_DIR="/home/we/.local/share/SuperCollider/Extensions"
+
+# Check if SuperCollider source directory exists
+if [ ! -d "$SC_DIR" ]; then
+    echo "ERROR: SuperCollider source directory not found at $SC_DIR."
+    exit 1
+fi
 
 # Create build directory
 cd $PROJ_DIR || 'exit'
@@ -26,22 +38,36 @@ echo "Making build directory... "
 mkdir build
 cd build || 'exit'
 
+# Start build process
+echo "Starting build from $(pwd)"
+cmake .. -DCMAKE_BUILD_TYPE='Release' -DSC_PATH="${SC_DIR}" -DCMAKE_INSTALL_PREFIX="${INSTALL_DIR}"
+cmake --build . --config Release
+
+# Check if the build was successful
+if [ $? -ne 0 ]; then
+    echo "ERROR: Build failed. Exiting"
+    exit 1
+else
+    echo "Build completed successfully."
+fi
+
 # Delete existing extensions
 if [ -d "${INSTALL_DIR}/Open303SuperCollider" ]; then
     echo "Deleting existing Open303SuperCollider extension..."
     rm -rf "${INSTALL_DIR}/Open303SuperCollider"
 fi
 
-# Start build process
-echo "Starting build from $(pwd)"
-cmake .. -DCMAKE_BUILD_TYPE='Release' -DSC_PATH="${SC_DIR}" -DCMAKE_INSTALL_PREFIX="${INSTALL_DIR}"
-cmake --build . --config Release
-
 echo "Build complete. Installing to ${INSTALL_DIR}"
 cmake --build . --config Release --target install
 
+# Check if the install was successful
+if [ $? -ne 0 ]; then
+    echo "ERROR: Install failed. Exiting"
+    exit 1
+fi
+
 # Delete Supernova plugin version
-echo "Deleting Supernova plugin version..."
+echo "Deleting Supernova plugin version if present..."
 rm -rf "${INSTALL_DIR}/Open303SuperCollider/Open303/Open303_supernova.scx"
 
 echo "Plugin built and installed to ${INSTALL_DIR}"
